Token handling (#38)
* Format exception descriptions which contained unformatted interpolations

* Refactor out user pool base URL templating

…so it can be reused by token verification.

* Fix several weaknesses in token verification

JWTs are made of footguns.  Follow the recommendations in Cognito's
developer guide and avoid some pitfalls with the defaults of the JWT
library in use (jose).

  • Actually verify audience and issuer claims against expected values
    instead of comparing the values to themselves.

  • Require audience, issuer, and expiration claims to be present, as
    otherwise the JWT library in use will silently skip verification.

  • Check token_use claim matches after claim verification to avoid
    using unverified claims at all.

* Consistently set and verify tokens across assignment codepaths

Ensures that the token attributes are accessible, up-to-date, and valid.

* Add verify_tokens() method to manually (re-)check current tokens

It can be useful to call this method after creating a Cognito instance
where you've provided externally-remembered token values.

* Save the verified claims for the id and access token

Very handy to have these available so you can use the default claims
like "cognito:groups" and any custom claims.

Resolves #19.

* README: Correct two typos

* Switch to f-strings for interpolation

As requested during review. I avoided using f-strings initially because
I did not see them used elsewhere in the codebase.

Co-authored-by: Pascal Vizeli <pascal.vizeli@syshack.ch>

Co-authored-by: Pascal Vizeli <pascal.vizeli@syshack.ch>